Fix: Fix the sign in error and merge SignInActivity with MainActivity
