+++ app/build.gradle
@@ -9,7 +9,7 @@ android {
         versionCode 1
         versionName "1.0"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
-        manifestPlaceholders = [appAuthRedirectScheme: "cloud.artik.example.hellocloud://oauth2callback"]
+        manifestPlaceholders = [appAuthRedirectScheme: "com.example.user.tgifire://oauth2callback"]
     }
     buildTypes {
         release {
