+++ app/src/main/AndroidManifest.xml
@@ -2,6 +2,8 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.example.user.tgifire">
 
+    <uses-permission android:name="android.permission.INTERNET" />
+
     <application
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
@@ -12,14 +14,32 @@
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
+                <action android:name="com.example.user.tgifire.ARTIKCLOUD_AUTHORIZATION_RESPONSE"/>
             </intent-filter>
         </activity>
-        <activity android:name=".SignInActivity" />
+
         <activity android:name=".BuildingInfoActivity" />
         <activity android:name=".UserMainActivity" />
         <activity android:name=".AdminMainActivity"></activity>
+
+
+        <activity
+            android:name="net.openid.appauth.RedirectUriReceiverActivity">
+            <intent-filter>
+                <action android:name="android.intent.action.VIEW"/>
+                <category android:name="android.intent.category.DEFAULT"/>
+                <category android:name="android.intent.category.BROWSABLE"/>
+
+                <!--
+                 Make sure these two intent filter fields respects "AUTH REDIRECT URL"
+                 of your application set up at the developer.artik.cloud.
+                 For example, if REDIRECT URL is "cloud.artik.example.oauth://oauth2callback",
+                 the intent-filter fields looks like the following
+                -->
+                <data android:scheme="com.example.user.tgifire" android:host="oauth2callback"/>
+            </intent-filter>
+        </activity>
     </application>
 
 </manifest>
