--- app/src/main/java/com/example/user/tgifire/SignInActivity.java
@@ -1,149 +1,0 @@
-package com.example.user.tgifire;
-
-import android.app.PendingIntent;
-import android.content.Context;
-import android.content.Intent;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.customtabs.CustomTabsIntent;
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-import android.util.Log;
-import android.widget.Toast;
-
-import net.openid.appauth.AuthState;
-import net.openid.appauth.AuthorizationException;
-import net.openid.appauth.AuthorizationRequest;
-import net.openid.appauth.AuthorizationResponse;
-import net.openid.appauth.AuthorizationService;
-import net.openid.appauth.TokenResponse;
-
-import static com.example.user.tgifire.AuthHelper.INTENT_ARTIKCLOUD_AUTHORIZATION_RESPONSE;
-import static com.example.user.tgifire.AuthHelper.USED_INTENT;
-
-public class SignInActivity extends AppCompatActivity {
-    static final String TAG = "SignInActivity";
-
-    AuthorizationService mAuthorizationService;
-    AuthStateDAL mAuthStateDAL;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        Log.v(TAG, "::onCreate");
-
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_sign_in);
-
-        mAuthorizationService = new AuthorizationService(this);
-        mAuthStateDAL = new AuthStateDAL(this);
-
-        doAuth();
-    }
-
-    // File OAuth call with Authorization Code with PKCE method
-    // https://developer.artik.cloud/documentation/getting-started/authentication.html#authorization-code-with-pkce-method
-    private void doAuth() {
-        AuthorizationRequest authorizationRequest = AuthHelper.createAuthorizationRequest();
-
-        PendingIntent authorizationIntent = PendingIntent.getActivity(
-                this,
-                authorizationRequest.hashCode(),
-                new Intent(INTENT_ARTIKCLOUD_AUTHORIZATION_RESPONSE, null, this, SignInActivity.class),
-                0);
-
-        /* request sample with custom tabs */
-        CustomTabsIntent.Builder builder = new CustomTabsIntent.Builder();
-        CustomTabsIntent customTabsIntent = builder.build();
-
-        mAuthorizationService.performAuthorizationRequest(authorizationRequest, authorizationIntent, customTabsIntent);
-    }
-
-    @Override
-    protected void onStart() {
-        Log.d(TAG, "Entering onStart ...");
-        super.onStart();
-        checkIntent(getIntent());
-    }
-
-    @Override
-    protected void onNewIntent(Intent intent) {
-        checkIntent(intent);
-    }
-
-    private void checkIntent(@Nullable Intent intent) {
-
-        Log.d(TAG, "Entering checkIntent ...");
-        if (intent != null) {
-            String action = intent.getAction();
-            switch (action) {
-                case INTENT_ARTIKCLOUD_AUTHORIZATION_RESPONSE:
-                    Log.d(TAG, "checkIntent action = " + action
-                            + " intent.hasExtra(USED_INTENT) = " + intent.hasExtra(USED_INTENT));
-                    if (!intent.hasExtra(USED_INTENT)) {
-                        handleAuthorizationResponse(intent);
-                        intent.putExtra(USED_INTENT, true);
-                    }
-                    break;
-                default:
-                    Log.w(TAG, "checkIntent action = " + action);
-                    // do nothing
-            }
-        } else {
-            Log.w(TAG, "checkIntent intent is null!");
-        }
-    }
-
-    private void handleAuthorizationResponse(@NonNull Intent intent) {
-        AuthorizationResponse response = AuthorizationResponse.fromIntent(intent);
-        AuthorizationException error = AuthorizationException.fromIntent(intent);
-        Log.i(TAG, "Entering handleAuthorizationResponse with response from Intent = " + response.jsonSerialize().toString());
-
-        if (response != null) {
-
-            if (response.authorizationCode != null ) { // Authorization Code method: succeeded to get code
-
-                final AuthState authState = new AuthState(response, error);
-                Log.i(TAG, "Received code = " + response.authorizationCode + "\n make another call to get token ...");
-
-                // File 2nd call to get the token
-                mAuthorizationService.performTokenRequest(response.createTokenExchangeRequest(), new AuthorizationService.TokenResponseCallback() {
-                    @Override
-                    public void onTokenRequestCompleted(@Nullable TokenResponse tokenResponse, @Nullable AuthorizationException exception) {
-                        if (tokenResponse != null) {
-                            authState.update(tokenResponse, exception);
-                            mAuthStateDAL.writeAuthState(authState); //store into persistent storage for use later
-                            String text = String.format("Received token response [%s]", tokenResponse.jsonSerializeString());
-                            Log.i(TAG, text);
-                            startBuildingInfoActivity();
-                        } else {
-                            Context context = getApplicationContext();
-                            Log.w(TAG, "Token Exchange failed", exception);
-                            CharSequence text = "Token Exchange failed";
-                            int duration = Toast.LENGTH_LONG;
-                            Toast toast = Toast.makeText(context, text, duration);
-                            toast.show();
-                        }
-                    }
-                });
-            } else { // come here w/o authorization code. For example, signup finish and user clicks "Back to login"
-                Log.d(TAG, "additionalParameter = " + response.additionalParameters.toString());
-
-                if (response.additionalParameters.get("status").equalsIgnoreCase("login_request")) {
-                    // ARTIK Cloud instructs the app to display a sign-in form
-                    doAuth();
-                } else {
-                    Log.d(TAG, response.jsonSerialize().toString());
-                }
-            }
-
-        } else {
-            Log.w(TAG, "Authorization Response is null ");
-            Log.d(TAG, "Authorization Exception = " + error);
-        }
-    }
-
-    private void startBuildingInfoActivity() {
-        Intent msgActivityIntent = new Intent(this, BuildingInfoActivity.class);
-        startActivity(msgActivityIntent);
-    }
-}
